pptx_viewer/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
python-pptx==0.6.21
